+++ app/src/Controller/TagController.php
@@ -146,6 +146,15 @@ class TagController extends AbstractController
     #[Route('/{id}/delete', name: 'tag_delete', requirements: ['id' => '[1-9]\d*'], methods: 'GET|DELETE')]
     public function delete(Request $request, Tag $tag): Response
     {
+        if (!$this->tagService->canBeDeleted($tag)) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.tag_contains_tasks')
+            );
+
+            return $this->redirectToRoute('tag_index');
+        }
+
         $form = $this->createForm(FormType::class, $tag, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('tag_delete', ['id' => $tag->getId()]),
